Final Project Submission (Mendoza&Jia)
This repository has the Electrospinning Project submission. It has three separate jupyter notebook with code and the final paper. Make sure you have the .csv file in the same directory as the jupyter notebook files. There are also several papers that were used as references for this project and please make sure you download Pandas and Plotly libraries. Specific instructions inside the document.